Refactor display_import functions to static methods
