+++ server_import.php
@@ -6,6 +6,7 @@
  * @package PhpMyAdmin
  */
 use PhpMyAdmin\Config\PageSettings;
+use PhpMyAdmin\Display\Import;
 use PhpMyAdmin\Response;
 
 /**
@@ -25,10 +26,9 @@ $scripts->addFile('import.js');
  */
 require 'libraries/server_common.inc.php';
 
-require 'libraries/display_import.lib.php';
 $response = Response::getInstance();
 $response->addHTML(
-    PMA_getImportDisplay(
+    Import::getImportDisplay(
         'server', $db, $table, $max_upload_size
     )
 );
